Fixes that are not related to rendering

+++ src/main/java/ru/bulldog/justmap/util/colors/ColorProviders.java
@@ -10,7 +10,7 @@ import net.minecraft.block.Blocks;
 import net.minecraft.block.RedstoneWireBlock;
 import net.minecraft.block.StemBlock;
 import net.minecraft.util.math.BlockPos;
-import net.minecraft.util.math.Vec3f;
+import net.minecraft.util.math.Vec3d;
 import net.minecraft.world.World;
 import net.minecraft.world.biome.Biome;
 import net.minecraft.world.chunk.Chunk;
@@ -50,8 +50,8 @@ public class ColorProviders implements ColorProvider {
 		}, Blocks.WATER, Blocks.BUBBLE_COLUMN, Blocks.CAULDRON);
 		blockColors.registerColorProvider((state, world, pos) -> {
 			int power = state.get(RedstoneWireBlock.POWER);
-			Vec3f powerVector = RedstoneLevelAccessor.getPowerVectors()[power];
-			return MathUtil.packRgb(powerVector.getX(), powerVector.getY(), powerVector.getZ());
+			Vec3d powerVector = RedstoneLevelAccessor.getPowerVectors()[power];
+			return MathUtil.packRgb((float) powerVector.getX(), (float) powerVector.getY(), (float) powerVector.getZ());
 		}, Blocks.REDSTONE_WIRE);
 		blockColors.registerColorProvider((state, world, pos) -> {
 			return Colors.ATTACHED_STEM;
